Remove unnecessary hooks
