+++ packages/react/src/progress/index.parts.ts
@@ -3,4 +3,4 @@ export { ProgressTrack as Track } from './track/ProgressTrack';
 export { ProgressIndicator as Indicator } from './indicator/ProgressIndicator';
 export { ProgressValue as Value } from './value/ProgressValue';
 
-export type { ProgressStatus as Status } from './root/useProgressRoot';
+export type { ProgressStatus as Status } from './root/ProgressRoot';
