+++ packages/react/src/progress/indicator/ProgressIndicator.test.tsx
@@ -2,29 +2,13 @@ import * as React from 'react';
 import { expect } from 'chai';
 import { Progress } from '@base-ui-components/react/progress';
 import { createRenderer, describeConformance, isJSDOM } from '#test-utils';
-import { ProgressRootContext } from '../root/ProgressRootContext';
-
-const contextValue: ProgressRootContext = {
-  max: 100,
-  min: 0,
-  value: 30,
-  formattedValue: '30',
-  status: 'progressing',
-  state: {
-    max: 100,
-    min: 0,
-    status: 'progressing',
-  },
-};
 
 describe('<Progress.Indicator />', () => {
   const { render } = createRenderer();
 
   describeConformance(<Progress.Indicator />, () => ({
     render: (node) => {
-      return render(
-        <ProgressRootContext.Provider value={contextValue}>{node}</ProgressRootContext.Provider>,
-      );
+      return render(<Progress.Root value={40}>{node}</Progress.Root>);
     },
     refInstanceof: window.HTMLDivElement,
   }));
