--- packages/react/src/progress/indicator/useProgressIndicator.ts
@@ -1,67 +1,0 @@
-'use client';
-import * as React from 'react';
-import { mergeProps } from '../../merge-props';
-import { valueToPercent } from '../../utils/valueToPercent';
-
-function useProgressIndicator(
-  parameters: useProgressIndicator.Parameters,
-): useProgressIndicator.ReturnValue {
-  const { max = 100, min = 0, value } = parameters;
-
-  const percentageValue =
-    Number.isFinite(value) && value !== null ? valueToPercent(value, min, max) : null;
-
-  const getStyles = React.useCallback(() => {
-    if (percentageValue == null) {
-      return {};
-    }
-
-    return {
-      insetInlineStart: 0,
-      height: 'inherit',
-      width: `${percentageValue}%`,
-    };
-  }, [percentageValue]);
-
-  const getRootProps: useProgressIndicator.ReturnValue['getRootProps'] = React.useCallback(
-    (externalProps = {}) =>
-      mergeProps<'span'>(
-        {
-          style: getStyles(),
-        },
-        externalProps,
-      ),
-    [getStyles],
-  );
-
-  return {
-    getRootProps,
-  };
-}
-
-namespace useProgressIndicator {
-  export interface Parameters {
-    /**
-     * The maximum value
-     * @default 100
-     */
-    max?: number;
-    /**
-     * The minimum value
-     * @default 0
-     */
-    min?: number;
-    /**
-     * The current value. The component is indeterminate when value is `null`.
-     */
-    value: number | null;
-  }
-
-  export interface ReturnValue {
-    getRootProps: (
-      externalProps?: React.ComponentPropsWithRef<'span'>,
-    ) => React.ComponentPropsWithRef<'span'>;
-  }
-}
-
-export { useProgressIndicator };
