+++ packages/react/src/progress/root/ProgressRoot.tsx
@@ -1,11 +1,33 @@
 'use client';
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import { formatNumber } from '../../utils/formatNumber';
+import { mergeProps } from '../../merge-props';
 import { useComponentRenderer } from '../../utils/useComponentRenderer';
-import { ProgressStatus, useProgressRoot } from './useProgressRoot';
+import { useLatestRef } from '../../utils/useLatestRef';
 import { ProgressRootContext } from './ProgressRootContext';
 import { progressStyleHookMapping } from './styleHooks';
 import { BaseUIComponentProps } from '../../utils/types';
+
+function formatValue(value: number | null, format?: Intl.NumberFormatOptions): string {
+  if (value == null) {
+    return '';
+  }
+
+  if (!format) {
+    return formatNumber(value / 100, [], { style: 'percent' });
+  }
+
+  return formatNumber(value, [], format);
+}
+
+function getDefaultAriaValueText(formattedValue: string | null, value: number | null) {
+  if (value == null) {
+    return 'indeterminate progress';
+  }
+
+  return formattedValue || `${value}%`;
+}
 
 /**
  * Groups all parts of the progress bar and provides the task completion status to screen readers.
@@ -32,42 +54,72 @@ const ProgressRoot = React.forwardRef(function ProgressRoot(
     ...otherProps
   } = props;
 
-  const { getRootProps, ...progress } = useProgressRoot({
-    'aria-label': ariaLabel,
-    'aria-labelledby': ariaLabelledby,
-    'aria-valuetext': ariaValuetext,
-    format,
-    getAriaLabel,
-    getAriaValueText,
-    max,
-    min,
-    value,
-  });
+  const formatOptionsRef = useLatestRef(format);
+
+  let status: ProgressStatus = 'indeterminate';
+  if (Number.isFinite(value)) {
+    status = value === max ? 'complete' : 'progressing';
+  }
+  const formattedValue = formatValue(value, formatOptionsRef.current);
 
   const state: ProgressRoot.State = React.useMemo(
     () => ({
       max,
       min,
-      status: progress.status,
+      status,
     }),
-    [max, min, progress.status],
+    [max, min, status],
   );
 
   const contextValue: ProgressRootContext = React.useMemo(
     () => ({
-      ...progress,
+      max,
+      min,
+      value,
+      formattedValue,
+      status,
       state,
     }),
-    [progress, state],
+    [max, min, value, formattedValue, status, state],
+  );
+
+  const propGetter = React.useCallback(
+    (externalProps = {}) =>
+      mergeProps<'div'>(
+        {
+          'aria-label': getAriaLabel ? getAriaLabel(value) : ariaLabel,
+          'aria-labelledby': ariaLabelledby,
+          'aria-valuemax': max,
+          'aria-valuemin': min,
+          'aria-valuenow': value ?? undefined,
+          'aria-valuetext': getAriaValueText
+            ? getAriaValueText(formattedValue, value)
+            : (ariaValuetext ?? getDefaultAriaValueText(formattedValue, value)),
+          role: 'progressbar',
+        },
+        otherProps,
+        externalProps,
+      ),
+    [
+      ariaLabel,
+      ariaLabelledby,
+      ariaValuetext,
+      formattedValue,
+      getAriaLabel,
+      getAriaValueText,
+      max,
+      min,
+      otherProps,
+      value,
+    ],
   );
 
   const { renderElement } = useComponentRenderer({
-    propGetter: getRootProps,
+    propGetter,
     render: render ?? 'div',
     state,
     className,
     ref: forwardedRef,
-    extraProps: otherProps,
     customStyleHookMapping: progressStyleHookMapping,
   });
 
@@ -78,6 +130,8 @@ const ProgressRoot = React.forwardRef(function ProgressRoot(
   );
 });
 
+export type ProgressStatus = 'indeterminate' | 'progressing' | 'complete';
+
 namespace ProgressRoot {
   export type State = {
     max: number;
@@ -85,8 +139,55 @@ namespace ProgressRoot {
     status: ProgressStatus;
   };
 
-  export interface Props extends useProgressRoot.Parameters, BaseUIComponentProps<'div', State> {}
+  export interface Props extends BaseUIComponentProps<'div', State> {
+    /**
+     * The label for the Indicator component.
+     */
+    'aria-label'?: string;
+    /**
+     * An id or space-separated list of ids of elements that label the Indicator component.
+     */
+    'aria-labelledby'?: string;
+    /**
+     * A string value that provides a human-readable text alternative for the current value of the progress indicator.
+     */
+    'aria-valuetext'?: string;
+    /**
+     * Options to format the value.
+     */
+    format?: Intl.NumberFormatOptions;
+    /**
+     * Accepts a function which returns a string value that provides an accessible name for the Indicator component.
+     * @param {number | null} value The component's value.
+     * @returns {string}
+     */
+    getAriaLabel?: (index: number | null) => string;
+    /**
+     * Accepts a function which returns a string value that provides a human-readable text alternative for the current value of the progress indicator.
+     * @param {string} formattedValue The component's formatted value.
+     * @param {number | null} value The component's numerical value.
+     * @returns {string}
+     */
+    getAriaValueText?: (formattedValue: string | null, value: number | null) => string;
+    /**
+     * The maximum value.
+     * @default 100
+     */
+    max?: number;
+    /**
+     * The minimum value.
+     * @default 0
+     */
+    min?: number;
+    /**
+     * The current value. The component is indeterminate when value is `null`.
+     * @default null
+     */
+    value: number | null;
+  }
 }
+
+export { ProgressRoot };
 
 ProgressRoot.propTypes /* remove-proptypes */ = {
   // ┌────────────────────────────── Warning ──────────────────────────────┐
@@ -172,5 +273,3 @@ ProgressRoot.propTypes /* remove-proptypes */ = {
    */
   value: PropTypes.number,
 } as any;
-
-export { ProgressRoot };
