+++ packages/react/src/progress/root/ProgressRootContext.tsx
@@ -1,9 +1,25 @@
 'use client';
 import * as React from 'react';
-import type { ProgressRoot } from './ProgressRoot';
-import type { useProgressRoot } from './useProgressRoot';
+import type { ProgressRoot, ProgressStatus } from './ProgressRoot';
 
-export type ProgressRootContext = Omit<useProgressRoot.ReturnValue, 'getRootProps'> & {
+export type ProgressRootContext = {
+  /**
+   * The maximum value.
+   */
+  max: number;
+  /**
+   * The minimum value.
+   */
+  min: number;
+  /**
+   * Value of the component.
+   */
+  value: number | null;
+  /**
+   * Formatted value of the component.
+   */
+  formattedValue: string;
+  status: ProgressStatus;
   state: ProgressRoot.State;
 };
