+++ packages/react/src/progress/value/ProgressValue.test.tsx
@@ -3,29 +3,13 @@ import { expect } from 'chai';
 import { spy } from 'sinon';
 import { Progress } from '@base-ui-components/react/progress';
 import { createRenderer, describeConformance } from '#test-utils';
-import { ProgressRootContext } from '../root/ProgressRootContext';
-
-const contextValue: ProgressRootContext = {
-  max: 100,
-  min: 0,
-  value: 30,
-  formattedValue: '30',
-  status: 'progressing',
-  state: {
-    max: 100,
-    min: 0,
-    status: 'progressing',
-  },
-};
 
 describe('<Progress.Value />', () => {
   const { render } = createRenderer();
 
   describeConformance(<Progress.Value />, () => ({
     render: (node) => {
-      return render(
-        <ProgressRootContext.Provider value={contextValue}>{node}</ProgressRootContext.Provider>,
-      );
+      return render(<Progress.Root value={40}>{node}</Progress.Root>);
     },
     refInstanceof: window.HTMLSpanElement,
   }));
